backend/requirements.txt
fastapi
uvicorn
python-dotenv
redis
cryptography
requests
python-dateutil
